# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cosco API Documentation",
        "description": "API documentation for Cosco application",
        "contact": {
            "email": "admin@cosco.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/memberships": {
            "get": {
                "tags": [
                    "Memberships"
                ],
                "summary": "Get user's memberships",
                "description": "Returns a list of all memberships for the authenticated user",
                "operationId": "getUserMemberships",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MembershipResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/memberships/status": {
            "get": {
                "tags": [
                    "Memberships"
                ],
                "summary": "Get user's active membership status",
                "description": "Returns the active membership status for the authenticated user",
                "operationId": "getUserMembershipStatus",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "has_active_membership": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "شما عضویت فعال دارید"
                                                },
                                                "remaining_days": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "membership": {
                                                    "$ref": "#/components/schemas/MembershipResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "has_active_membership": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "شما عضویت فعالی ندارید"
                                                },
                                                "remaining_days": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "membership": {
                                                    "type": "null"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "has_active_membership": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "عضویت شما منقضی شده است"
                                                },
                                                "remaining_days": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "membership": {
                                                    "$ref": "#/components/schemas/MembershipResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/memberships/{id}": {
            "get": {
                "tags": [
                    "Memberships"
                ],
                "summary": "Get a specific membership",
                "description": "Returns details of a specific membership (only user's own memberships)",
                "operationId": "getMembership",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Membership ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MembershipResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not authorized to view this membership"
                    },
                    "404": {
                        "description": "Membership not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/membership-types/{id}": {
            "get": {
                "tags": [
                    "Membership Types"
                ],
                "summary": "Get a specific membership type",
                "description": "Returns details of a specific membership type",
                "operationId": "getMembershipType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Membership type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MembershipTypeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Membership type not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MembershipResource": {
                "title": "Membership Resource",
                "description": "Membership resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "serial_number": {
                        "type": "string",
                        "example": "MEM-001-2024"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-01"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "status_badge": {
                        "type": "string",
                        "example": "فعال"
                    },
                    "remaining_days": {
                        "type": "integer",
                        "example": 15
                    },
                    "duration": {
                        "type": "string",
                        "example": "1 month"
                    },
                    "membership_type": {
                        "$ref": "#/components/schemas/MembershipTypeResource"
                    }
                },
                "type": "object"
            },
            "MembershipTypeResource": {
                "title": "Membership Type Resource",
                "description": "Membership type resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Basic Membership"
                    },
                    "description": {
                        "type": "string",
                        "example": "Basic membership with standard benefits"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 50000
                    },
                    "formatted_price": {
                        "type": "string",
                        "example": "50,000 ریال"
                    },
                    "day_cycle": {
                        "type": "integer",
                        "example": 30
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}